refactor: migrate createGridCardBlockSchema to modules
