+++ packages/core/client/src/schema-initializer/items/RecordAssociationGridCardBlockInitializer.tsx
@@ -4,7 +4,8 @@ import React, { useCallback } from 'react';
 import { useCollectionManager_deprecated } from '../../collection-manager';
 import { SchemaInitializerItem, useSchemaInitializer, useSchemaInitializerItem } from '../../application';
 import { useSchemaTemplateManager } from '../../schema-templates';
-import { createGridCardBlockSchema, useRecordCollectionDataSourceItems } from '../utils';
+import { useRecordCollectionDataSourceItems } from '../utils';
+import { createGridCardBlockSchema } from '../../modules/blocks/data-blocks/grid-card/createGridCardBlockSchema';
 
 /**
  * @deprecated
@@ -31,11 +32,9 @@ export const RecordAssociationGridCardBlockInitializer = () => {
           insert(
             createGridCardBlockSchema({
               rowKey: collection.filterTargetKey,
-              collection: field.target,
-              resource,
+              collectionName: field.target,
               dataSource: collection.dataSource,
               association: resource,
-              settings: 'blockSettings:gridCard',
             }),
           );
         }
@@ -57,10 +56,9 @@ export function useCreateAssociationGridCardBlock() {
       insert(
         createGridCardBlockSchema({
           rowKey: collection.filterTargetKey,
-          collection: field.target,
+          collectionName: field.target,
           dataSource: collection.dataSource,
           association: `${field.collectionName}.${field.name}`,
-          settings: 'blockSettings:gridCard',
         }),
       );
     },
